### 2021-04-17 #3183 @kyleniemeyer
Hi @JeanBilheux, thanks for your interest in JOSS. The editors will now discuss if it meets the [research software criterion](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) for review by JOSS. You should hear back in a week or so.
