### 2021-05-07 #3183 @kthyng
Hi @JeanBilheux! The editorial board has decided that this submission is in scope for us, though the argument for this being research software should be improved in your paper, as part of the review process. So, I have removed the "query-scope" label. Unfortunately, we have a backlog of submissions currently in this area with overburdened editors so I will add you to the backlog with the "waitlist" label. Thanks for your patience.
