### 2021-05-10 #3183 @JeanBilheux
Thanks @kthyng , I will improve the paper.md document and explain in more detail why this library is so important for the type of research we do.
