### 2021-06-28 #3426 @whedon
[REVIEW]: Circular profile of a ring over a stack of images
**Submitting author:** @jeanbilheux (<a href="http://orcid.org/0000-0003-2172-6487">Jean Bilheux</a>)
**Repository:** <a href="https://github.com/neutronimaging/circular_profiler" target ="_blank">https://github.com/neutronimaging/circular_profiler</a>
**Version:** v1.0.0
