### 2021-06-28 #3426 @jgostick
Hello @michaelberks and @prasantapal, the review is now under way.  You can find a checklist at the top of this issue thread, one for each of you.  Your task is simply to check off all those boxes.  This mostly involves installing the software according to the provided documentation, and verifying it's functionality.  If you find any boxes that you *cannot* check, then you let the author(s) of the package know and they should address it.  You can ping them through this issue, or by opening an issue on their github repo directxly.

Thank you both very much for your time.
