### 2021-08-09 neutronimaging/circular_profiler#2 @michaelberks
Error during environment set-up
I am installing as part of JOSS review https://github.com/openjournals/joss-reviews/issues/3426.

I followed the installation instructions in README.md, however when running config_conda_env.sh I received an installation error regarding nodejs:

> ValueError: Please install nodejs >=12.0.0 before continuing. nodejs may be installed using conda or directly from the nodejs website.

I tried manually installing nodejs into the condo env and then re-running the install script but the same error occurred. The attached file shows the full output of the script the second time it was run.

[install_output.txt](https://github.com/neutronimaging/circular_profiler/files/6953167/install_output.txt)

I tried running the notebook anyway, but the GUI wouldn't launch correctly, I presume because of some missing dependency due to the installation error.
