### 2021-08-09 #3426 @michaelberks
@JeanBilheux my initial review comments *in italics*:

## Functionality

###  Installation:
Does installation proceed as outlined in the documentation?
*I have tried installing the software but received an error running the condo environment setup script. I have logged a new issue in the source code repo. I am unable to proceed with reviewing functionality.*

## Documentation

###  Example usage:
Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
*The software would benefit from including a small test dataset of images, with specific description of how to run the GUI tool on this data*

###  Functionality documentation:
Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
*There is minimal documentation, however I get the impression the tool is supposed to be simple and intuitive to use once the GUI is running. Because I could not get the GUI to run (see installation error above) I cannot confirm if this is the case though.*

###  Automated tests:
Are there automated tests or manual steps described so that the functionality of the software can be verified?
*There is a tests folder in the repo, and the last commit comment mentions starting to implement unit tests, but it is not clear if this is complete, and no instructions are given for running the tests*

###  Community guidelines:
Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
*I could not find these guidelines*

## Software paper

###  A statement of need:
Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
*I think the statement of need could be clearer in how this applies more widely to the target audience. At first read it appears to solve a very niche problem, which may only be applicable to the authors' own lab*

###  State of the field:
Do the authors describe how this software compares to other commonly-used packages?
*No*

###  Quality of writing:
Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
*The summary is very clear. I think the **statement of need** requires some work though. It isn't clear in sentence 2 who the users are or what the sample is - a real physical object or the data files from sentence 1. Then in sentence 3, what is the large volume of data? Is all of this just referring the fact that the output of a CT scan is a set of image slices (stored in TIFF or FITS format)? What specific type of analysis is time consuming? What type of symmetry is required of the sample? The proposed software achieves a very specific type of analysis, and it isn't clear what specific properties of the sample being imaged need to be to make this software applicable. It is then mentioned this has been used on equipment in the authors' lab, to analyse a specific fault. It isn't clear to me from the rest of the *statement of need* that this software generalises to other problems. By that I mean, clearly code that computes intensity profiles in a ring across a stack of images can be applied to **any** set of images, but it isn't clear why these measurements would be useful in other datasets.*

###  References:
Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
*No references are given, this again makes me slightly concerned about the generalisability of the software to other existing problems, or whether it is of use to other researches*
