### 2021-08-31 fffe175
Update config_conda_env.sh
