### 2021-08-31 neutronimaging/circular_profiler#2 @JeanBilheux
I was unable to reproduce the error but got this error in the past that I fixed by giving the exact version of nodejs needed to be able to run the notebook. I updated the `config_conda_env.sh` file.
