### 2021-09-01 f41d190
added unit test files and updated launcher script. this refs #3
