+++ launch_notebook.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+jupyter nbextension enable --py widgetsnbextension
+jupyter notebook circular_profile_of_a_ring.ipynb
