### 2021-09-01 #3426 @JeanBilheux
**Functionality documentation**

I added a quick tutorial in the README but a more detail instruction can be found [here](https://neutronimaging.pages.ornl.gov/tutorial/notebooks/circular_profile_of_a_ring/#activate-search) as mentioned at the top of the README.
