### 2021-09-01 917d6c8
Fixed most of unit tests warning. this refs #4
