+++ src/calculate_profiles.py
@@ -115,8 +115,8 @@ class CalculateProfiles:
         full_angles_matrix = np.zeros((image_height, image_width))
 
         # bottom right corner of matrix
-        right_bottom_corner_width = np.int(image_width - x_central_pixel)
-        right_bottom_corner_height = np.int(image_height - y_central_pixel)
+        right_bottom_corner_width = int(image_width - x_central_pixel)
+        right_bottom_corner_height = int(image_height - y_central_pixel)
 
         x_right_bottom = np.arange(right_bottom_corner_width) + x_central_pixel
         y_right_bottom = np.arange(right_bottom_corner_height) + y_central_pixel
@@ -126,13 +126,13 @@ class CalculateProfiles:
         angles_right_bottom = 90 + np.rad2deg(np.arctan((yv_right_bottom - y_central_pixel) /
                                                         (xv_right_bottom - x_central_pixel)))
 
-        full_angles_matrix[np.int(y_right_bottom[0]): np.int(y_right_bottom[0]) + len(y_right_bottom),
-                           np.int(x_right_bottom[0]): np.int(x_right_bottom[0]) + len(x_right_bottom)] = \
+        full_angles_matrix[int(y_right_bottom[0]): int(y_right_bottom[0]) + len(y_right_bottom),
+                           int(x_right_bottom[0]): int(x_right_bottom[0]) + len(x_right_bottom)] = \
             angles_right_bottom
 
         # top right corner of matrix
-        right_top_corner_width = np.int(image_width - x_central_pixel)
-        right_top_corner_height = np.int(y_central_pixel)
+        right_top_corner_width = int(image_width - x_central_pixel)
+        right_top_corner_height = int(y_central_pixel)
 
         x_right_top = np.arange(right_top_corner_width) + x_central_pixel
         y_right_top = np.arange(right_top_corner_height)
@@ -142,12 +142,12 @@ class CalculateProfiles:
         angles_right_top = 90 + np.rad2deg(np.arctan((yv_right_top - y_central_pixel) /
                                                      (xv_right_top - x_central_pixel)))
 
-        full_angles_matrix[np.int(y_right_top[0]): np.int(y_right_top[0])+len(y_right_top),
-                           np.int(x_right_top[0]):np.int(x_right_top[0])+len(x_right_top)] = angles_right_top
+        full_angles_matrix[int(y_right_top[0]): int(y_right_top[0])+len(y_right_top),
+                           int(x_right_top[0]):int(x_right_top[0])+len(x_right_top)] = angles_right_top
 
         # top left corner
-        left_top_corner_width = np.int(x_central_pixel)
-        # left_top_corner_height = np.int(y_central_pixel)
+        left_top_corner_width = int(x_central_pixel)
+        # left_top_corner_height = int(y_central_pixel)
 
         x_left_top = np.arange(left_top_corner_width)
         y_left_top = np.arange(y_central_pixel)
@@ -156,12 +156,12 @@ class CalculateProfiles:
 
         angles_left_top = 270 + np.rad2deg(np.arctan((yv_left_top - y_central_pixel) /
                                                      (xv_left_top - x_central_pixel)))
-        full_angles_matrix[np.int(y_left_top[0]): np.int(y_left_top[0])+len(y_left_top),
-                           np.int(x_left_top[0]): np.int(x_left_top[0])+len(x_left_top)] = angles_left_top
+        full_angles_matrix[int(y_left_top[0]): int(y_left_top[0])+len(y_left_top),
+                           int(x_left_top[0]): int(x_left_top[0])+len(x_left_top)] = angles_left_top
 
         # bottom left corner
[…]
