### 2021-09-01 9c7b31c
how to run the tests. this refs #4
