### 2021-09-01 7091fcd
update unit test instructions. this refs #4
