### 2021-09-01 #3426 @JeanBilheux
**Automated tests**

I added instructions on how to run the unit tests and fixed the important warnings.
https://github.com/neutronimaging/circular_profiler/issues/4
