### 2021-09-01 #3426 @JeanBilheux
I added some comments about **Community Guidelines** in the README.
