### 2021-09-01 #3426 @JeanBilheux
**Statement of need**

Rewrote it by using a more general example where this tool can be used.
