### 2021-11-12 #3426 @jgostick
Hi @JeanBilheux..firstly, I am very sorry about how long this has taken! I have dedicated this afternoon to doing the review of your package, and have run into some issues during installation and usage:

* I use windows 10, so installed your package using the WSL2 tool.  It took ages, which I'm sure would also be the case on a native linux machine.  The problem was conda trying to resolve the conflicts.  I used Python 3.8.8, because I wanted to see if I could ignore the 3.7 you put in the instructions.  It eventually installed, but perhaps this caused the conflicts?
* Once I opened the jupyter notebook, I was unable to run it. First it told me I did not have ``qtpy``, but installing that only led to a new error about "No Qt bindings could be found".  So this led me to install qt using: "sudo apt-get install qt5-default", which took another long while.
* That led to the same error, so I figured perhaps I need PySide, so I did "sudo apt-get install python3-pyside".  This did not help.  I am now officially stuck.  I do not know how to install the package.  It seems like your installation instructions are missing some dependencies.
* It was also quite unusual to clone the repo in order to run the code.  I expect to clone a repo only if I actually plan to edit the code.  For this use case I would expect something like ``conda install circular_profiler``.  Additionally, running the shell script was unusual.  The jupyter notebook that get launched does not have a lot of code in it.  Could you not pivot to telling the users to create their own empty notebook, then provide them with instructions on on how to use your package to get the profiler running?  Like if I'm using scikit-image, I open an empty notebook, then start typing some commands until I get the result I want.  Could your package be interacted with as a library like this?
