### 2021-11-12 #3426 @jgostick
Since I was unable to run it, I thought I would take a look at some of other checkboxes:

* tests.  The Readme has a badge that says the build is passing, which I found surprizing given how many dependencies this package has, so I clicked on the badge and got a 404 from TravisCI: https://www.travis-ci.com/neutronimaging/circular_profiler.  This is not a link to a branch specific build, this is the main travis page for the repo.  So it seems like you've migrated away from Travis as the CI?  Not sure why the badge is green though...seems like Travis should return something like a grey badge for a missing repo.
* need.  GUIs are nice, but are they always necessary? Could I not get this profile from just analyzing the stack of images using numpy functions?  I spent about an hour on this, just because I like coding more than doing real work, and came up with the following.  There are only 14 lines of actual code.  I know this is not as handy as your gui, and probably slower too, but I "think" it offers the same functionality. So I'm left wondering how the LOC counter found 1800 lines. Presumably all from the Qt interface?  Getting back to the issue of need, was a gui so badly needed that it justified 100x more work than necessary, not to mention the extreme challenges of the installation.  In fact, there is a bit of irony here...your package is basically only accessible to highly computer literate people (linux users), and this is the exact audience that would have no problems writing a small python script like mine below.  You could even package the logic in a function to make it a single call for the user.  Basically, the need is lacking.

``` python
import numpy as np
import matplotlib.pyplot as plt
from tqdm import tqdm

# %% User input
im = np.random.rand(500, 500, 500) > 0.8  # The image to be analyzed

# User specified ring properties
c = [240, 250]  # center
ri, ro = (100, 150)  # inner and outer radii
n = 50  # Number of angular bins to use

# %% The code
# Generate a ring
xx, yy = np.meshgrid(range(im.shape[0]), range(im.shape[1]))
d = np.sqrt((xx - c[0])**2 + (yy - c[1])**2)
ring = (d > ri) * (d < ro)
# Make it a 3D annulus
annulus = np.tile(np.atleast_3d(ring), [1, 1, im.shape[2]])
# Create theta field for thresholding below
q = np.rad2deg(np.arctan((yy - c[1])/(xx - c[0])))
q[xx < c[0]] = q[xx < c[0]] + 180
q += 90

# Scan image for specified number of angles
bins = np.linspace(0, 360, n, endpoint=False)
vals = np.zeros_like(bins)
data = im*annulus
w = 360/n
for i, a in tqdm(enumerate(bins)):
    mask = (q > a)*(q < (a+w))*annulus
    vals[i] = data[mask].sum()/data[mask].size

# %%  Do some plotting
fig, ax = plt.subplots(1, 4)
ax[0].imshow(im[..., 50], origin='lower')
ax[1].imshow(ring, origin='lower')
ax[2].imshow(data[..., 50], origin='lower')
ax[3].plot(bins, vals, 'b-o')
ax[3].plot([0, 360], [np.mean(vals)]*2)
plt.ylim([0, 1])
```

* scholarly effort. Relating to my above point, the effort that went into this package does not seem 'scholarly' to me.  I'm sure the gui took effort, but the scholarly, technical, researchy part seems very small. JOSS does not like publishing papers that are only about GUIs.
* paper.  I'm not at all happy with the paper.md.  This is a cut and paste of the readme and does not really even try to be a useful readable paper for future users.  It does include one additional piece of info regarding the 'need' (ie neutron images of turbine blades), and whlie I agree that we need to study this, it does not tell me why users need this software. Is this feature not available in imagej or dragonfly or avizo, etc?  Is it really necessary to interact with the center and size of the rings in real time?
* community guidelines: This only contains a single issue which you wish someone to implement.  This should be more general, like "If you wish to contribute to this package, here are the detailed instructions about how we do things..."  and "You can find suitable issues to work on by looking at our bug tracker", etc.
