### 2021-11-15 #3426 @bidochon
* I'm so sorry about the build that did not work. I tested it on Linux and mac and didn't have access to a windows machine at that time (I do now). So not being able to build the program is my mistake. Sorry about that.
* This program is part of a big suite of tools (that I have been implemented for our users) and that I isolated for the purpose of this publication.
* This program, and the entire suite of tools, is available to our users when they connect to our server where all their experiment data seat. This way they have nothing to install, just 2 clicks and they have the entire suite of tools available (as described in the tutorial I provided in the publication -> neutronimaging.pages.ornl.gov)
* I didn't know JOSS was aiming at just algorithms and didn't like UI. that's good to know.
* Like many other languages, Python has many ways to come up with the same logic. I implemented mine, the reviewer implemented his. Yes, this is a minor part of the entire code but I regret that the reviewer is not familiar with user interface implementation as those always require much more lines of codes than the algorithm they usually wrap up around (and make user friendly). Instead of the UI I implemented I could have just provided a notebook with the line of codes (such as the one the reviewer implemented here) but I would probably (surely) be out of a job right now. I'm not going to debate about all the types of users we have but having a user interface is a must for the type of research and users we serve.
* I have been trying to describe the necessity of this program for our users but apparently failed.

So for now, I'm going to retract this publication for the following reasons:
* I failed at providing an easy to install version for all OS
* I failed at explaining the interest of this algorithm, and why a UI is a MUST to have to our users
* JOSS does not seem to like UI, but only algorithms.
