### 2021-11-15 #3426 @jgostick
I am sorry that I chased you away.  I am a JOSS associate editor, but I do not speak for JOSS officially...I just sheppard along the review process (in your case by acting as the second reviewer).  However, JOSS editors tend to decline GUI-only type contributions (like the Shiny apps written for R) during the "scope check" phase (which your package passed).  These usually represent wrappers around other code, and don't contribute much/any core functionality.  When I think about whether a package is of appropriate "scope" I usually ask myself, "would a user of this software feel the need to cite it in their actual research paper".   To me this means "does this software package do something fundamental enough that an author feels the need to point the reader to a source explaining its function".  By this definition, a GUI-only software is not sufficient.  Instead an author would point to the paper describing the underlying functionality that the GUI wraps since that's the code that actually touched/manipluated/produced the data.  Am I making sense?

In your case, you DO offer some core functionality, so it did make it past the scopy check.  However, I was curious how much functionality was actually required to accomplish this, so wrote the above function (also for fun).  After this excerise I am not not 100% convinced that your software offers sufficient functionality.  On the other hand, the mission of JOSS is to provide software developers with a means of getting recognition for their time and efforts.  I am wiling to continue this review if you are still intersted.  A "quick and effortless install on all possible platforms" is not a requirement for JOSS so I can try to install it on the Ubuntu box in my lab.  I had hoped that using WSL2 would work (it does for a lot things, even OpenFOAM), but I guess your GUI requirements implicitly depend on some of the core components (probably related to the desktop UI?).

All that being said, I feel like your submission still requires quite a lot of work. Many of the 'check boxes' are not satisfied and overall your submission feels incomplete.  So, I think that a temporary withdrawal of the submission makes sense, and you can resubmit later once things are more polished.

Here are some suggestions:

* You mention that this is part a large package of tools, so maybe the whole package could be submitted?
* Most packages submitted to JOSS have a "brand" or catchy name, so perhaps you need to think about how you market this.  Instead of "Circular profile of a ring over a stack of images", perhaps the title of your submission could be "RingThing: A python-based GUI for obtaining radial intensity profiles in volumetric images".
* If you decide to submit the whole package, then you could do "NeutronStare: A toolkit for analzying neutron intensity images"  (I hope this one made you chuckle at least a little bit!)
* The paper is not meant to be an exhaustive deepdive into the code, but it needs to include some sort of presentable text.  For instance, some images and output of the code, along with some sort of explanation of how the function works.
* The web-presence of a package is important too, so the name should searchable and recognizable.  If you're really serious about offer this package to the world, then maybe spend $10 per year on a domain name?
